Move client protocol to the client realm
